Fixed bug when there is no previous log, added some logging

+++ harness/utils/logging.py
@@ -3,7 +3,8 @@ import os
 
 
 def setup_logging(logfile):
-    os.remove(logfile)
+    if os.path.exists(logfile):
+        os.remove(logfile)
     logging.basicConfig(filename=logfile, format='%(asctime)s %(message)s', level=logging.DEBUG)
